Rebase the OpenBSD Anoubis trees onto a current OpenBSD
The OpenBSD Anoubis tree is rebased onto a the current OpenBSD tree.
This should fix the problems caused by the incorrect conversion from
OpenBSD to Git. In order to make the tree build a few very small fixes
had to be made after the rebase. The necessary changes to the build
environment to support this are also part of this change. Finally the
new kernel requires a new vmstat binary which is added to the binaries
that are being built with the kernel. The only test that uses this
binary is adjusted accordingly.

Aegis-Change: anoubis.10.600.C337 Delta 79
Bug: 314
BLE