Fix memory leak in __anoubis_event_common() when no event queue is
defined

Currently, it is up to __anoubis_event_common() to arrange the release
of the memory it receives in 'buf' as an argument. The function does so
by calling eventdev_enqueue_*wait(), which will free the memory at some
point (immediately, or when appropriate). However, __anoubis_event_
common() can return before calling these functions, in case no event
queue is defined. In this case, it needs to release the memory
explicitly.

Aegis-Change: anoubis.10.750.C514 Delta 43
TE BLE