Avoid deadlock on vnodes
Currently we call getcwd from a mac hook while we hold the vnode lock
on a vnode that is being opened. getcwd will lock each directory vnode
on the path to the root (one node is locked at a time). This can dead
lock with another process that is trying do a lookup because during a
lookup both the directory and the directory entry are locked at the
same time and the directory lock is taken first. Temporary solution is
to drop the lock on the leaf node during the call to getcwd.

Aegis-Change: anoubis.10.810.C176 Delta 10
Bug: 690
TE BLE