Protect network interfaces' MAC hooks with IPL_NET in
OpenBSD

As found out "the hard way" by Stefan Fritsch, the locking currently
done around the MAC hooks dealing with network interfaces is incorrect.
We can also be called from interrupt context, in which case grabbing a
process-to-process lock is not only wrong but impossible, resulting in
the need of additional protection through the use of appropriate
splnet() and splx() calls to block interrupts coming up to the IPL_NET
(architecture-dependent kernel defined value) priority.

Aegis-Change: anoubis.10.830.C146 Delta 18
Bug: 720
BLE