allow programs to delegate checksum calculation to kernel
We introduce a new ioctl(2) on /dev/anoubis. This ioctl allows a
program to specify an open file descriptor. The kernel will return the
checksum for this descriptor if it can be calculated. Currently only
anoubisctl makes use of the new ioctl. In the long run we should get
rid of the ANOUBIS_CHECKSUM_OP_CALC anbd replace ANOUBIS_CHECKSUM_OP_
ADD with a variant that includes the desired checksum in the request.

Aegis-Change: anoubis.10.900.C129 Delta 33
Bug: 771
BLE