Fix NFS crashes
The new security_dentry_open hook is called at a point where the file
is almost but not quite completely open. For this reason reading data
from the file for checksum calculation is possible on most but not all
filesystems. One of the exceptions is NFS. Solution: Move the call to
the dentry_open hook down until after the call to the low level open
function.

Aegis-Change: anoubis.10.900.C172 Delta 42
Bug: 710
TE BLE