: Fix 'NULL mp in getnewvnode()' warning message
If we are coming from getdevvp() (meaning we are allocating a vnode for
a block or character device), it is normal to have 'mp' set to NULL.
This debug code comes from FreeBSD, which has devfs. We don't, so the
correct fix in our case is to remove the printf() entirely.

Aegis-Change: anoubis.10.950.C337 Delta 120
BLE