: Fix alf/KernelPolicy.pm, ManageQemuInstances.pm and
XenTest.pm after sysctl change to OpenBSD

Change 373 modified the name of the 'anoubis' sysctl to 'mac_anoubis'.
Therefore, the tests (as well as the userland sysctl tool) need to be
adapted accordingly. This change also fixes the declaration of the
ANOUBIS_NAMES array in sysctl.h -- it is supposed to be composed of
<sysctl name, sysctl type> pairs, and not <sysctl name, sysctl number>
pairs.

Aegis-Change: anoubis.10.950.C391 Delta 160
BLE