Generate unlock events on Linux
To correctly detect the end of package upgrades we need to detect
package database unlock events. This can be done via the security_file_
lock hook. Along the way we fix a bug in fs/locks.c which causes flock-
style arguments to be passed to the security hook instead of fcntl
ones.

Aegis-Change: anoubis.10.970.C620 Delta 278
BLE