Fix EBADF error returned from dhclient on fedora
Fedora dhclient receives an EBADF when writing to the raw socket. This
is because getpeername in the kernel does not return a peer for an
unconnected raw socket. Do not retrun EBADF in this case if the address
familty is not INET/INET6. Instead ask the daemon and proceed according
to the result. The daemon should not expected suitable addresses in an
event in this case anyway because it does not know about the address
family.

Aegis-Change: anoubis.10.971.C587 Delta 258
TE BLE