Try to fix OpenBSD nfs_reply crash
Fix an OpenBSD bug that can cause crashes in nfs_reply: nfs_receive is
expected to either return data in *mp or an error. Returning no error
but no data either, i.e. EOF is not allowed because this would imply
that an nfs reply is missing. NOTE: The very same fix is in netbsd as
well. Got fixed there in 1999.

Aegis-Change: anoubis.10.971.C670 Delta 311
BLE