Fix OpenBSD pool allocator bug
Calls to pool_do_get are protected by the pool's mutex. If pool_do_get
decides that it has to sleep that mutex is dropped as it is not allowed
to schedule while holding a mutex. However, this logic does not take
the case into accout where pool_get/pool_do_get is called recursively.
This can happen for the internal page header pool (phpool). If this
inner call to pool_get sleeps it will only drop the mutex of the inner
pool. Fix this by not allowing that call to sleep.

Aegis-Change: anoubis.10.972.C235 Delta 72
BLE