Fix NFS related OpenBSD panic
The memory handling for NFS requests is partially flawed. Fix. Some
more background: The NFS code sets the SAVENAME flag on the struct
component during nfs_lookup (aka VFS_LOOKUP) if the lookup mode is not
'LOOKUP' (but 'DELETE, CREATE etc.). This means that a) the path
component is not freed by the namei code and b) the NFS code is
responsible for freeing the data itself. However, if the SAVESTART flag
is given by the caller, the name belongs to the caller instead. Even
VOP_ABORTOP will not try to free it. Thus the NFS code should in theory
not free the path name buffer if SAVESTART is set in cn_flags. This is
a theoretical bug in NFS code that cannot be triggered in plain OpenBSD
because this situation never happens. The modified lookup flags used in
Anoubis expose this bug and we need to fix all the call sites where the
NFS code frees the namei buffer.

Aegis-Change: anoubis.10.972.C407 Delta 181
BLE