NULL pointer dereference in linux kernel
If the sfs module is loaded between the start end the end of an exec it
is possible that sfs_bprm_committed_creds gets called with a task that
does not have a security label attached to it. Make sure that we do not
panic in this case!

Aegis-Change: anoubis.10.972.C476 Delta 222
Bug: 1447
BLE