Playground-ID and associated ioctl
Implement an ioctl that allows a task to open itself in a new
playground environment. This is a relatively small change that does not
yet include tests. However, this should enable someone to write a
suitable client client utility that is required for tests. Extensive
tests will be added after the initial client utility and the read only
lookup is implemented. The sandbox/DaemonPolicy.pm test was run but is
not part of the test because if conflicts with Change 642.

Aegis-Change: anoubis.10.972.C654 Delta 340
Bug: 1497
TE BLE