Add infrastructure for playground module
This change adds infrastructure for the playground module to linux
kernel. This module will deal with inode labels etc. It defines the
name of the security attribute for playground labels and makes sure
that the security attribute cannot be written by normal users while the
playground module is active. No other functionality is added at this
time. TEST: sandbox/DaemonPolicy.pm passed but is in conflict with
another change.

Aegis-Change: anoubis.10.972.C689 Delta 357
Bug: 1504
TE BLE