Move playground code to playground.c
This change moves playground related code from anobuis_core.c to
playground.c. This is the clean way to do it and avoids lots of ugly
ifdefs. There are a view functions in fs/namei.c that could move to
playground.c. too. We don't do this because theses functions can be
inlined if fs/namei.c.

Aegis-Change: anoubis.10.972.C707 Delta 372
Bug: 1504
BLE