Support write operations in the playground
Ensure that write operations (i.e. open of regular files for writing)
copy the file into the playground. To the reviewer: A git-tree that
contains the commits that correspond to this change is available at /
scratch/ehrhardt/pg2/linux-git/. Just use git-diff on the old and new
tree in the GITINFO file to see a better versions of the modifications
done in this change.

Aegis-Change: anoubis.10.972.C726 Delta 383
Bug: 1510
BLE