Support write operations in the playground
Bugfixes and an extensive test. Bugfixes include:
* O_APPEND does not imply open for write. Doing so causes spurious
errors on open
* An O_EXCL flag withouth O_CREAT must not lead to EEXIST. This could
happen if the playground added an implicit O_CREAT.
* If O_CREAT is not given and the file must be copied, take the new
permissions from the original file. Make sure that this is propagated
to the caller and that the value is initialized to zero if this should
ever fail.

Aegis-Change: anoubis.10.972.C732 Delta 386
Bug: 1510
BLE