Support link/rename in the playground
This is the next batch of playground write changes. It includes
additional test, bugfixes and intial (manually tested) support for
rename and hardlinks. Rename works only partially, because the old file
is not turned into a whiteout yet. What follows is a more detailed list
of changes:
* Security Hooks have been added that limit link, rename unlink and
rmdir operations of the playground.
* The do_lookup code has been simplified siginificantly. Operations
that create a new file can never take this path.
* Lookups for the source of a rename or hardlink now return an error
code in the playground if the source is not yet a playground file. The
caller must initiate a copy operation into the playground and then
restart the system call.
* The code has been compile tested with and without CONFIG_SECURITY_
ANOUBIS.
* The behaviour of open was bogus for some cases if the name given to
open is a symbolic link. This has been fixed and tests have been added
that test for these cases.

Aegis-Change: anoubis.10.972.C744 Delta 387
Bug: 1510
BLE