Playground copy: Support symlink cloning
This change adds support for clonging of symbolic links in case of
hardlinks and rename. List of all modifications:
* Pass MAY_ACCESS down to inode_permission and return ok if sys_access
asks for write permission on a production system file (fixes need for -
f flag to rm)
* Bail out earlier during lookup if neither playground nor original
file exits.
* Change error codes to make it clearer for callers what needs to be
done. Rename the copy function for files and add a new copy function
for symlinks. The new error codes can be used to find out which of
these functions must be called.
* The symlink clone function needs a hack that allows it to create a
symklink in the playground if the symlink the the production system
already exits
* Add an implementation for the readlink security hook. Symlinks can be
used to sneak data out of a playground environment via symlink/
readlink.
* Apparently, hardlinks from production to playground files were
broken. This is now fixed and some tests need adjustment.

Aegis-Change: anoubis.10.972.C764 Delta 401
Bug: 1511
BLE