Do not allow production files in playground dirs
In playground directories it is neccessary that files in that directory
can be removed. This is done e.g. while cleaning up the playground.
However, it should not be possible to create new object within a
playground directory unless they belong to the same playground.
Implement appropriate hooks.

Aegis-Change: anoubis.10.972.C914 Delta 494
BLE