Fix runaway modproble loop when loading sha256 early
Loading the crypto modules early (e.g. in sfs_init) can lead to console
message before the console is initialized. This triggers a modprobe for
the console device which fails an in turn tries to write to the
console. This results in an endless loop.

Aegis-Change: anoubis.10.972.C920 Delta 497
TE BLE