Make playground.c dazukofs aware
Unfortunately, the playground module must be aware of the dazukofs
module in order to properly support playgrounds on dazukofs. This still
requires that the dazukofs mount is on the same mount point as the base
directory for now. This change basically ensures that playground ID and
all other playground information is maintained on the lower inode not
on the upper inode. This is neccessary to avoid bugs that are caused
because the dazukofs module does not call security hooks before calling
function on the lower fs. An additional change is a bugfix to dazukofs
which can cause memory leaks in dazukofs if mkdir/rmdir is called in a
loop. This leak leads to missing inode delete events, too.

Aegis-Change: anoubis.10.973.C182 Delta 70
BLE