Support voluntary playground renames and use in X wrapper script
This change allows playground commands to declare themselves as the
main playground process that should show up in playground lists. The
real command name is still shown in parenthesis. This is used in the
xpgwrapper script to display a better playground name to the user.
Additionally, the xpgwrapper script is changed to start the client
program from within the icewm window manager session and not from the
outside. This avoids serveral problems including playground sessions
that do not terminate due to remaining dbus processes. The old code
that tries to kill playground processes based on anoubisctl ps is
removed as it does not work if a certificate is configured.

Aegis-Change: anoubis.10.974.C194 Delta 50